yolo_server/requirements.txt
flask
pillow
numpy
ultralytics
gunicorn